Change data type in RecyclerDialogAdapter and apply Stable Id
- 작업을 제대로 삭제하기 위해서는 Task의 Id을 사용하는게 좋다고 판단하여 RecyclerDialogAdapter에서 todoList가 아닌 TaskList를 가지도록 변경
- RecyclerDialogAdapter 성능을 올리기 위하여 StableId 추가

+++ app/src/main/java/com/example/making_calendar/data/database/Task.kt
@@ -11,5 +11,5 @@ data class Task (
     var time: String?,
     var todo: String,
 ){
-    @PrimaryKey(autoGenerate = true) var id: Int = 0;
+    @PrimaryKey(autoGenerate = true) var id: Long = 0;
 }
